pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["DariGlad"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
envparse = "^0.2.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-filter = "^22.1"
drf-spectacular = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
